Ignore snapshot@1
snapshot@1 should be granted as being preapplied to the state.

A node's log is empty at boot but having a snapshot@1 is strange.
Since the log's snapshot index is 0, there should be no snapshot of
the future clock. After inserting the snapshot entry@1,
the application should have snapshot@1 but it can be nothing.