fix: skip snapshot insertion
It is possible that the same snapshot is sent from the leader.
In this case, the follower can skip the insertion.

I don't think this is a logical bug because leader can send
a newer snapshot after it creates one.
However, rejecting replication impedes the consensus process of the
cluster which is harmful.